requirements.txt
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
python-binance==1.0.19
PyYAML==6.0.1
python-telegram-bot==20.5
python-dateutil==2.8.2
